refactor: reorganize TaskDialogCard and TaskModal imports; move hooks to local paths and remove unused hook files
